+++ pkg/models/plan.go
@@ -20,10 +20,10 @@ type Vaktplan struct {
 
 // GuardDuty keeps track of minutes not worked in a given guard duty
 type GuardDuty struct {
-	Hvilende2000                 int
-	Hvilende0006                 int
-	Hvilende0620                 int
-	Helgetillegg                 int
-	Skifttillegg                 int
+	Hvilende2000                 float64
+	Hvilende0006                 float64
+	Hvilende0620                 float64
+	Helgetillegg                 float64
+	Skifttillegg                 float64
 	WeekendOrHolidayCompensation bool
 }
